Authentication Applied

+++ src/app/task.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Task } from './Task.interface';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
 @Injectable({
@@ -12,22 +12,22 @@ export class TaskService {
   constructor(private httpClient: HttpClient) {}
 
   getAllTasks(): Observable<Task[]> {
-    return this.httpClient.get<Task[]>(this.baseUrl);
+    return this.httpClient.get<Task[]>(this.baseUrl,  { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   getTask(id: string): Observable<Task> {
-    return this.httpClient.get<Task>(this.baseUrl + '/' + id);
+    return this.httpClient.get<Task>(this.baseUrl + '/' + id, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   addTask(payload: Task) {
-    return this.httpClient.post(this.baseUrl, { ...payload, status: 'TODO' });
+    return this.httpClient.post(this.baseUrl, { ...payload, status: 'TODO' }, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   deleteTask(id: string) {
-    return this.httpClient.delete(this.baseUrl + '/' + id);
+    return this.httpClient.delete(this.baseUrl + '/' + id, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   updateTask(id: string, updatedTask: Task) {
-    return this.httpClient.put(this.baseUrl + '/' + id, updatedTask);
+    return this.httpClient.put(this.baseUrl + '/' + id, updatedTask, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 }
